apurisk/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-docx>=1.1.0
jinja2>=3.1.2
pyyaml>=6.0
matplotlib>=3.7.0
